+++ platform/api/azure/api.go
@@ -18,140 +18,124 @@ import (
 	"context"
 	"fmt"
 	"math/rand"
+	"net/url"
 	"os"
-	"os/user"
-	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/classic/management"
-	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2022-08-01/compute"
-	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2021-02-01/network"
-	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2020-10-01/resources"
-	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2021-01-01/subscriptions"
-	armStorage "github.com/Azure/azure-sdk-for-go/services/storage/mgmt/2021-01-01/storage"
-	"github.com/Azure/azure-sdk-for-go/storage"
-	"github.com/Azure/go-autorest/autorest"
-	"github.com/Azure/go-autorest/autorest/azure/auth"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v5"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/network/armnetwork/v5"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armresources"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/resources/armsubscriptions"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/storage/armstorage"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
 	"github.com/coreos/pkg/capnslog"
-
-	internalAuth "github.com/flatcar/mantle/auth"
 )
 
 var (
 	plog = capnslog.NewPackageLogger("github.com/flatcar/mantle", "platform/api/azure")
 )
 
+type credData struct {
+}
+
 type API struct {
-	client      management.Client
-	rgClient    resources.GroupsClient
-	depClient   resources.DeploymentsClient
-	imgClient   compute.ImagesClient
-	compClient  compute.VirtualMachinesClient
-	vmImgClient compute.VirtualMachineImagesClient
-	netClient   network.VirtualNetworksClient
-	subClient   network.SubnetsClient
-	ipClient    network.PublicIPAddressesClient
-	intClient   network.InterfacesClient
-	accClient   armStorage.AccountsClient
+	cloudConfig cloud.Configuration
+	creds       azcore.TokenCredential
+	subID       string
+
+	rgClient    *armresources.ResourceGroupsClient
+	depClient   *armresources.DeploymentsClient
+	imgClient   *armcompute.ImagesClient
+	compClient  *armcompute.VirtualMachinesClient
+	vmImgClient *armcompute.VirtualMachineImagesClient
+	netClient   *armnetwork.VirtualNetworksClient
+	subClient   *armnetwork.SubnetsClient
+	ipClient    *armnetwork.PublicIPAddressesClient
+	intClient   *armnetwork.InterfacesClient
+	accClient   *armstorage.AccountsClient
 	Opts        *Options
 }
 
 type Network struct {
-	subnet network.Subnet
-}
-
-func setOptsFromProfile(opts *Options) error {
-	profiles, err := internalAuth.ReadAzureProfile(opts.AzureProfile)
-	if err != nil {
-		return fmt.Errorf("couldn't read Azure profile: %v", err)
-	}
-
-	if os.Getenv("AZURE_AUTH_LOCATION") == "" {
-		if opts.AzureAuthLocation == "" {
-			user, err := user.Current()
-			if err != nil {
-				return err
-			}
-			opts.AzureAuthLocation = filepath.Join(user.HomeDir, internalAuth.AzureAuthPath)
-		}
-		// TODO: Move to Flight once built to allow proper unsetting
-		os.Setenv("AZURE_AUTH_LOCATION", opts.AzureAuthLocation)
-	}
-
-	var subOpts *internalAuth.Options
-	if opts.AzureSubscription == "" {
-		settings, err := auth.GetSettingsFromFile()
-		if err != nil {
-			return err
-		}
-		subOpts = profiles.SubscriptionOptions(internalAuth.FilterByID(settings.GetSubscriptionID()))
-	} else {
-		subOpts = profiles.SubscriptionOptions(internalAuth.FilterByName(opts.AzureSubscription))
-	}
-	if subOpts == nil {
-		return fmt.Errorf("Azure subscription named %q doesn't exist in %q", opts.AzureSubscription, opts.AzureProfile)
-	}
-
-	if opts.SubscriptionID == "" {
-		opts.SubscriptionID = subOpts.SubscriptionID
-	}
-
-	if opts.SubscriptionName == "" {
-		opts.SubscriptionName = subOpts.SubscriptionName
-	}
-
-	if opts.ManagementURL == "" {
-		opts.ManagementURL = subOpts.ManagementURL
-	}
-
-	if opts.ManagementCertificate == nil {
-		opts.ManagementCertificate = subOpts.ManagementCertificate
-	}
-
-	if opts.StorageEndpointSuffix == "" {
-		opts.StorageEndpointSuffix = subOpts.StorageEndpointSuffix
-	}
-
-	return nil
+	subnet armnetwork.Subnet
 }
 
 // New creates a new Azure client. If no publish settings file is provided or
 // can't be parsed, an anonymous client is created.
 func New(opts *Options) (*API, error) {
-	var err error
-	conf := management.DefaultConfig()
-	conf.APIVersion = "2015-04-01"
+	var (
+		err    error
+		creds  azcore.TokenCredential
+		subID  string
+		config cloud.Configuration
+	)
 
-	if opts.ManagementURL != "" {
-		conf.ManagementURL = opts.ManagementURL
+	disableInstanceDiscovery := strToBool(os.Getenv("AZURE_DISABLE_INSTANCE_DISCOVERY"), false)
+	if opts.ADHost != "" {
+		config.ActiveDirectoryAuthorityHost = opts.ADHost
+	} else {
+		if opts.CloudName == "" {
+			opts.CloudName = os.Getenv("AZURE_CLOUD")
+		}
+		config, err = getCloudConfiguration(opts.CloudName)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get cloud config: %w", err)
+		}
+	}
+
+	authOpts := azidentity.DefaultAzureCredentialOptions{
+		ClientOptions: azcore.ClientOptions{
+			Cloud: config,
+		},
+		// Not setting the AdditionallyAllowedTenants here, use
+		// AZURE_ADDITIONALLY_ALLOWED_TENANTS env var to set up
+		// extra tenants - the azidentity module will take care
+		// of it.
+		AdditionallyAllowedTenants: nil,
+		DisableInstanceDiscovery:   disableInstanceDiscovery,
+		TenantID:                   os.Getenv("AZURE_TENANT_ID"),
+	}
+	creds, err = azidentity.NewDefaultAzureCredential(&authOpts)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get default Azure credentials: %w", err)
+	}
+
+	if opts.PreferredSubscriptionID == "" {
+		opts.PreferredSubscriptionID = os.Getenv("AZURE_SUBSCRIPTION_ID")
+	}
+
+	subIDs, err := querySubscriptions(context.TODO(), creds, config)
+	if err != nil {
+		return nil, fmt.Errorf("failed to query available subscriptions: %w", err)
+	}
+	if len(subIDs) == 0 {
+		return nil, fmt.Errorf("no subscriptions are available for default credentials")
+	}
+	if opts.PreferredSubscriptionID == "" {
+		if len(subIDs) > 1 {
+			return nil, fmt.Errorf("many available subscriptions, need to specify a preferred subscription (e.g. through AZURE_SUBSCRIPTION_ID env var)")
+		}
+		subID = subIDs[0]
+	} else {
+		for _, queried := range subIDs {
+			if queried == opts.PreferredSubscriptionID {
+				subID = opts.PreferredSubscriptionID
+				break
+			}
+		}
+		if subID == "" {
+			return nil, fmt.Errorf("preferred subscription %s is not a part of available subscriptions", opts.PreferredSubscriptionID)
+		}
 	}
 
 	if opts.StorageEndpointSuffix == "" {
-		opts.StorageEndpointSuffix = storage.DefaultBaseURL
-	}
-
-	if !opts.UseIdentity {
-		err = setOptsFromProfile(opts)
-		if err != nil {
-			return nil, fmt.Errorf("failed to get options from azure profile: %w", err)
-		}
-	} else {
-		subid, err := msiGetSubscriptionID()
-		if err != nil {
-			return nil, fmt.Errorf("failed to query subscription id: %w", err)
-		}
-		opts.SubscriptionID = subid
-	}
-
-	var client management.Client
-	if opts.ManagementCertificate != nil {
-		client, err = management.NewClientFromConfig(opts.SubscriptionID, opts.ManagementCertificate, conf)
-		if err != nil {
-			return nil, fmt.Errorf("failed to create azure client: %v", err)
-		}
-	} else {
-		client = management.NewAnonymousClient()
+		opts.StorageEndpointSuffix = "core.windows.net"
 	}
 
 	if opts.AvailabilitySet != "" && opts.ResourceGroup == "" {
@@ -159,8 +143,10 @@ func New(opts *Options) (*API, error) {
 	}
 
 	api := &API{
-		client: client,
-		Opts:   opts,
+		cloudConfig: config,
+		creds:       creds,
+		subID:       subID,
+		Opts:        opts,
 	}
 
 	err = api.resolveImage()
@@ -171,98 +157,110 @@ func New(opts *Options) (*API, error) {
 	return api, nil
 }
 
-func (a *API) newAuthorizer(baseURI string) (autorest.Authorizer, error) {
-	if !a.Opts.UseIdentity {
-		return auth.NewAuthorizerFromFile(baseURI)
+func (a *API) SetupClients() error {
+	opts := &arm.ClientOptions{
+		ClientOptions: policy.ClientOptions{
+			Cloud: a.cloudConfig,
+		},
 	}
-	settings, err := auth.GetSettingsFromEnvironment()
+
+	rcf, err := armresources.NewClientFactory(a.subID, a.creds, opts)
+	if err != nil {
+		return err
+	}
+	a.rgClient = rcf.NewResourceGroupsClient()
+	a.depClient = rcf.NewDeploymentsClient()
+
+	ccf, err := armcompute.NewClientFactory(a.subID, a.creds, opts)
+	if err != nil {
+		return err
+	}
+	a.imgClient = ccf.NewImagesClient()
+	a.compClient = ccf.NewVirtualMachinesClient()
+	a.vmImgClient = ccf.NewVirtualMachineImagesClient()
+
+	ncf, err := armnetwork.NewClientFactory(a.subID, a.creds, opts)
+	if err != nil {
+		return err
+	}
+	a.netClient = ncf.NewVirtualNetworksClient()
+	a.subClient = ncf.NewSubnetsClient()
+	a.ipClient = ncf.NewPublicIPAddressesClient()
+	a.intClient = ncf.NewInterfacesClient()
+
+	scf, err := armstorage.NewClientFactory(a.subID, a.creds, opts)
+	if err != nil {
+		return err
+	}
+	a.accClient = scf.NewAccountsClient()
+
+	return nil
+}
+
+func (a *API) GetBlobServiceClient(storageAccount string) (*service.Client, error) {
+	accountURL := fmt.Sprintf("https://%s.blob.%s", url.PathEscape(storageAccount), url.PathEscape(a.Opts.StorageEndpointSuffix))
+	if _, err := url.Parse(accountURL); err != nil {
+		return nil, err
+	}
+	opts := &service.ClientOptions{
+		ClientOptions: azcore.ClientOptions{
+			Cloud: a.cloudConfig,
+		},
+	}
+	return service.NewClient(accountURL, a.creds, opts)
+}
+
+func strToBool(v string, onInvalid bool) bool {
+	switch v {
+	case "yes", "y", "true", "t", "1":
+		return true
+	case "no", "n", "false", "f", "0":
+		return false
+	default:
+		return onInvalid
+	}
+}
+
+func getCloudConfiguration(name string) (cloud.Configuration, error) {
+	switch name {
+	case "", "public", "pub":
+		return cloud.AzurePublic, nil
+	case "china", "cn":
+		return cloud.AzureChina, nil
+	case "government", "gov":
+		return cloud.AzureGovernment, nil
+	default:
+		return cloud.Configuration{}, fmt.Errorf("invalid Azure cloud name: %s", name)
+	}
+}
+
+func querySubscriptions(ctx context.Context, creds azcore.TokenCredential, config cloud.Configuration) ([]string, error) {
+	opts := &arm.ClientOptions{
+		ClientOptions: policy.ClientOptions{
+			Cloud: config,
+		},
+	}
+	client, err := armsubscriptions.NewClient(creds, opts)
 	if err != nil {
 		return nil, err
 	}
-	return settings.GetMSI().Authorizer()
-}
-
-func msiGetSubscriptionID() (string, error) {
-	settings, err := auth.GetSettingsFromEnvironment()
-	if err != nil {
-		return "", err
-	}
-	subid := settings.GetSubscriptionID()
-	if subid != "" {
-		return subid, nil
-	}
-	auther, err := settings.GetMSI().Authorizer()
-	if err != nil {
-		return "", err
-	}
-	client := subscriptions.NewClient()
-	client.Authorizer = auther
-	iter, err := client.ListComplete(context.Background())
-	if err != nil {
-		return "", fmt.Errorf("failed to list subscriptions: %w", err)
-	}
-	for sub := iter.Value(); iter.NotDone(); iter.Next() {
-		// this should never happen
-		if sub.SubscriptionID == nil {
-			continue
+	// ClientListOptions is an empty struct, so pass nil instead
+	pager := client.NewListPager(nil)
+	var subIDs []string
+	for pager.More() {
+		page, err := pager.NextPage(ctx)
+		if err != nil {
+			return nil, err
 		}
-		if subid != "" {
-			return "", fmt.Errorf("multiple subscriptions found; pass one explicitly using the %s environment variable", auth.SubscriptionID)
+		for _, sub := range page.Value {
+			// this should never happen
+			if sub.SubscriptionID == nil {
+				continue
+			}
+			subIDs = append(subIDs, *sub.SubscriptionID)
 		}
-		subid = *sub.SubscriptionID
 	}
-	if subid == "" {
-		return "", fmt.Errorf("no subscriptions found; pass one explicitly using the %s environment variable", auth.SubscriptionID)
-	}
-	plog.Infof("Using subscription %s", subid)
-	return subid, nil
-}
-
-func (a *API) SetupClients() error {
-	auther, err := a.newAuthorizer(resources.DefaultBaseURI)
-	if err != nil {
-		return err
-	}
-	subid := a.Opts.SubscriptionID
-
-	a.rgClient = resources.NewGroupsClient(subid)
-	a.rgClient.Authorizer = auther
-
-	a.depClient = resources.NewDeploymentsClient(subid)
-	a.depClient.Authorizer = auther
-
-	auther, err = a.newAuthorizer(compute.DefaultBaseURI)
-	if err != nil {
-		return err
-	}
-	a.imgClient = compute.NewImagesClient(subid)
-	a.imgClient.Authorizer = auther
-	a.compClient = compute.NewVirtualMachinesClient(subid)
-	a.compClient.Authorizer = auther
-	a.vmImgClient = compute.NewVirtualMachineImagesClient(subid)
-	a.vmImgClient.Authorizer = auther
-
-	auther, err = a.newAuthorizer(network.DefaultBaseURI)
-	if err != nil {
-		return err
-	}
-	a.netClient = network.NewVirtualNetworksClient(subid)
-	a.netClient.Authorizer = auther
-	a.subClient = network.NewSubnetsClient(subid)
-	a.subClient.Authorizer = auther
-	a.ipClient = network.NewPublicIPAddressesClient(subid)
-	a.ipClient.Authorizer = auther
-	a.intClient = network.NewInterfacesClient(subid)
-	a.intClient.Authorizer = auther
-
-	auther, err = a.newAuthorizer(armStorage.DefaultBaseURI)
-	if err != nil {
-		return err
-	}
-	a.accClient = armStorage.NewAccountsClient(subid)
-	a.accClient.Authorizer = auther
-
-	return nil
+	return subIDs, nil
 }
 
 func randomNameEx(prefix, separator string) string {
@@ -270,6 +268,7 @@ func randomNameEx(prefix, separator string) string {
 	rand.Read(b)
 	return fmt.Sprintf("%s%s%x", prefix, separator, b)
 }
+
 func randomName(prefix string) string {
 	return randomNameEx(prefix, "-")
 }
@@ -286,7 +285,7 @@ func (a *API) GC(gracePeriod time.Duration) error {
 		return fmt.Errorf("listing resource groups: %v", err)
 	}
 
-	for _, l := range *listGroups.Value {
+	for _, l := range listGroups {
 		if strings.HasPrefix(*l.Name, "kola-cluster") {
 			createdAt := *l.Tags["createdAt"]
 			timeCreated, err := time.Parse(time.RFC3339, createdAt)